Fix: Invalid String when Importing sus Chart

+++ Ched/Plugins/SusParser.cs
@@ -255,7 +255,7 @@ namespace Sus2Image.Converter
                     break;
 
                 case "WAVEOFFSET":
-                    WaveOffset = int.Parse(value);
+                    WaveOffset = double.Parse(value);
                     break;
 
                 case "JACKET":
